requirements.txt
flask>=3.0.0
anthropic>=0.25.0
reportlab>=4.1.0
cryptography>=42.0.0
discord.py>=2.3.0
python-dotenv>=1.0.0
spacy>=3.7.0
# After install, also run: python -m spacy download en_core_web_sm
